server/toia-dm/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
blis==0.7.5
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloud-sql-python-connector==0.5.1
cryptography==36.0.1
cymem==2.0.6
fastapi==0.74.0
frozenlist==1.3.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.0
matplotlib
multidict==6.0.2
murmurhash==1.0.6
nltk==3.7
numpy==1.22.2
openai
packaging==21.3
pandas==1.4.1
pathy==0.6.1
plotly
preshed==3.0.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
srsly==2.4.2
starlette==0.17.1
thinc==8.0.13
threadpoolctl==3.1.0
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
wasabi==0.9.0
yarl==1.7.2